+++ contracts/interfaces/IERC721FrameFactory.sol
@@ -19,6 +19,6 @@ interface IERC721FrameFactory {
     function dropNFT(
         string memory _name,
         string memory _symbol,
-        uint256 _royaltyPercentage
+        uint96 _royaltyPercentage // in basis points (1 bp = 0.01%)
     ) external;
 }
